Create LibraryCatVer2.36
Contains a functional Search History button which shows after the first search. If you want an initial button can be turned on inside the MainWindow init.  There is a problem though with the Search button in that it keeps re-making the button over top previous instances of the button creating a confusing interface.